[Editor] Frontend to use editor service API to retrieve connections during session creation (#48)
* Set up basic structure and define connection model

* Initialize structure for controller

* Add connection CRUD functionalities

* Formatting files

* Update port for editor

* Add temporary frontend for testing editor

* Remove hardcoded document key

* Delete document key in shareDB

* Beautify editor and add syntx highlighting

* Editor frontend to retrieve connection from service

+++ services/editor/api/connectionController.js
@@ -43,7 +43,7 @@ class ConnectionController {
         }
         if (!is_used) {
           var port = i;
-          var document_key = "document_" + port; 
+          var document_key = "document_" + port;
           // console.log(port);
           const new_connection = new Connection({
             port: port,
@@ -70,27 +70,56 @@ class ConnectionController {
   }
 
   delete(req, res) {
-    var delete_query = Connection.find({
-      session_id: req.params.session_id,
-    }).deleteMany();
-    delete_query.exec(function (err, result) {
+    var found = false;
+    var query = Connection.find({ session_id: req.params.session_id });
+    query.exec(function (err, connection) {
       if (err) {
         return res.status(400).json({
           status: "error",
           message: err,
         });
       }
-      if (result["deletedCount"] < 1) {
+      if (connection.length < 1) {
         return res.status(400).json({
           status: "error",
-          message: "Session ID not found!",
+          message: "No connections found!",
         });
       }
-      res.json({
-        message: "Connection deleted",
-        data: result,
-      });
+      var document_key = connection[0]["document_key"];
+      console.log(document_key);
+      try {
+        service.remove_document(document_key);
+        found = true;
+      } catch (err) {
+        res.status(400).json({
+          status: "error",
+          message: err,
+        });
+      }
     });
+    if (found) {
+      var delete_query = Connection.find({
+        session_id: req.params.session_id,
+      }).deleteMany();
+      delete_query.exec(function (err, result) {
+        if (err) {
+          return res.status(400).json({
+            status: "error",
+            message: err,
+          });
+        }
+        if (result["deletedCount"] < 1) {
+          return res.status(400).json({
+            status: "error",
+            message: "Session ID not found!",
+          });
+        }
+        res.json({
+          message: "Connection deleted",
+          data: result,
+        });
+      });
+    }
   }
 
   retrieve(req, res) {
